requirements.txt
Flask==2.3.2
Flask-Login==0.6.3
Flask-WTF>=1.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
email-validator==2.0.0
bcrypt==4.0.1
weasyprint==60.0
Pillow==10.0.0
WTForms==3.0.1
python-dotenv==1.0.0
pytest>=7.0.0